Escape </script and <!-- in inlined <script>s.
Scripts that contain the string "</script" close the <script> tag when
inlined. This occurs in real code (e.g., in polymer 0.5.1), so it needs
to be worked around.

This change escapes scripts that contain the string "</script" (and
variations w/ whitespace), using the same trick as polymer's vulcanize
tool (which borrowed it from UglifyJS).